src/SVC1/requirements.txt
x==0.115.6
fastapi-cli==0.0.7
mysql-connector-python==9.0.0
sqlmodel==0.0.22
pydantic==2.10.5
pydantic_core==2.27.2
python-dotenv==1.0.1
pylint==3.3.3